rag-chatbot-langchain/requirements.txt
langchain-google-genai
langchain-community
chromadb
python-dotenv
langchain